tests/unit/: Breaks DependencyRegister tests into public API tests and other unit tests.
This is a strategy to ensure we test for behaviour documented as part of the public API, and can also test for more "private" behaviour to achieve 100% test coverage.

This is essentially a balance of "black box" testing for the public API, and "white box" testing with knowledge of internal implementations details which may change.

The goal is that while implementation details may change, and the "white box" tests will change with them, the "public API" tests will ensure we continue to meet documented behaviour.
